ai_service_voucher_platform/requirements.txt
# --- 核心依赖 (用于AI模型和知识库) ---
langchain-community
langchain-core
sentence-transformers==3.0.1
faiss-cpu
pydantic==2.9.2
PyYAML==6.0.1
pypdf
requests==2.32.3

# --- 辅助工具依赖 ---
watchdog==4.0.1

# --- API服务依赖 (用于main_api.py) ---
fastapi
uvicorn[standard]
httpx  # 用于异步HTTP回调